Version Final

+++ test.cpp
@@ -11,8 +11,6 @@
 #include <setjmp.h>
 #include <sys/time.h>
 
-
-
 using namespace std;
 
 #define PS1             "Simple-Shell$ "
@@ -29,6 +27,7 @@ void exitShell(){
 
     exit(0);
 }
+
 struct command
 {
   char **argv;
@@ -36,17 +35,19 @@ struct command
 
 void  INThandler(int sig){
 
-     char  c;
+     char * c = (char*)malloc(sizeof(char)*2);
      signal(sig, SIG_IGN);
-     printf("\nDesea Salir de la Shell? [Y/n] ");
-     scanf("%c",&c);
+     printf("\nDesea Salir de la Shell? [y/n] ");
      cin.clear();
-
-     if (c == 'y' || c == 'Y'|| c == '\n'){
-         exitShell();
-     }      
-
+     cin.getline(c,2*sizeof(char));
+     cin.clear();
+  
+    if (c[0] == 'y' ){
+        exitShell();
+     }
+    free(c);
     signal(SIGINT, INThandler);
+    
     
 }
 
@@ -101,6 +102,7 @@ int commandExist(char** cmd, int create){// cmd[0] no nos sirve; cmd[1] nombre d
     fclose(archivo);
     return 0;
 }
+
 void execCommand(char** cmd){
     int rep = atoi(cmd[3])/atoi(cmd[2]);
     
@@ -118,7 +120,7 @@ void execCommand(char** cmd){
 }
 
 void launchProgram(char** cmd, int log) {
-    int pid = fork();
+    pid_t pid = fork();
     int fw;
     if(log==1){
         fw=open("pato.txt",O_APPEND|O_WRONLY);
@@ -145,7 +147,6 @@ void launchProgram(char** cmd, int log) {
     }
 
     if(log==1) close(fw);
-    //return;
 }
 
 void createCommand(char** cmd){
@@ -161,6 +162,7 @@ void createCommand(char** cmd){
     }
     fclose(archivo);
 }
+
 void doPipe(struct command *args,int npipe){
 
     if(npipe>2){
@@ -176,14 +178,17 @@ void doPipe(struct command *args,int npipe){
     pid_t pid3;
 
     if (pid == 0){
+
+        printf("\033[0;96m");
+        fflush(stdout);
         dup2(pipes[1], 1);// 1PIPE
         
         close(pipes[0]);  // 1PIPE
         close(pipes[1]);  // 1PIPE
 
-        if(npipe>1){
-            close(pipes[2]);  // 2PIPE
-            close(pipes[3]);  // 2PIPE
+        if(npipe>1){    // 2PIPE
+            close(pipes[2]); 
+            close(pipes[3]);  
         }
 
         execvp(args[0].argv[0],(char* const*)args[0].argv);
@@ -195,12 +200,10 @@ void doPipe(struct command *args,int npipe){
             dup2(pipes[0], 0);  // 1PIPE
             
 
-            if(npipe>1){
-
-                dup2(pipes[3], 1);// 2PIPE
-                
-                close(pipes[2]);  // 2PIPE
-                close(pipes[3]);  // 2PIPE
+            if(npipe>1){ // 2PIPE
+                dup2(pipes[3], 1);
+                close(pipes[2]);
+                close(pipes[3]); 
             }
 
             close(pipes[0]);  // 1PIPE
@@ -214,10 +217,10 @@ void doPipe(struct command *args,int npipe){
                 if (pid3 == 0){ // 2PIPE
                     dup2(pipes[2], 0);
                 
-                    close(pipes[2]);  // 2PIPE
-                    close(pipes[3]);  // 2PIPE
-                    close(pipes[0]);  // 1PIPE
-                    close(pipes[1]);  // 1PIPE
+                    close(pipes[2]);
+                    close(pipes[3]);
+                    close(pipes[0]);
+                    close(pipes[1]);
 
                     execvp(args[2].argv[0],(char* const*)args[2].argv);
                 }    
@@ -240,10 +243,10 @@ void doPipe(struct command *args,int npipe){
         wait(NULL);
         
     }
-    
+    printf("\033[0;93m");
     kill(pid,SIGKILL);
-    kill(pid2,SIGKILL);
-    if(npipe>1)kill(pid3,SIGKILL);
+    if(npipe>1)kill(pid2,SIGKILL);
+    //if(npipe>1)kill(pid3,SIGKILL);
     
   }
 
@@ -268,7 +271,6 @@ command* argsPipe(char * line, int n){
        token = strtok(auxLine, "|");
        if(token!=NULL){
             args2[i] = token;
-            //cout<<args2[i]<<endl; 
        }
        auxLine = NULL;    
     }
@@ -292,7 +294,6 @@ int main(){
         
         char* line = getCommandLine();
         int n = isPipe(line);
-        //cout<<n<<endl;
         int flag = 0;
         if(!n){
             char** cmd = splitLine(line);
@@ -313,18 +314,8 @@ int main(){
         }
         else{
             command * args = argsPipe(line,n);
-            
             doPipe(args,n);
-            
-
-            /*for(int i=0;i<n+1;i++){
-                for(int j=0;args[i].argv[j]!=NULL;j++){
-                    cout<<args[i].argv[j]<<" ";
-                }
-                cout<<endl;
-            }*/
         }
     }
-
     return 0;
 }
